feat: support build-packages
The cause of the problem is that the base Ubuntu image used by the
Snapcraft rock has an empty Apt cache. This makes sense for application
images which don't expect to use/need Apt, but is a problem for a build
tool like Snapcraft.

The fix in this commit is to add a "wrapper" script that first calls
"apt update" to populate the cache, and then execs Snapcraft as normal.