build: drop chisel from rockcraft.yaml (#17)
I don't recall *why* specifically I added chisel in the Rockcraft
project, but the Snapcraft rock should not *more* capable than regular
Snapcraft. This means using Chisel only if provided by Snapcraft itself,
and Snapcraft 7 does *not* support Chisel.

When merging this to core22-8 we'll re-add the Chisel spread test, but
in a v8-specific suite.